requirements.txt

flask==3.0.2
python-dotenv==1.0.1
groq==0.5.0
requests==2.31.0
gunicorn==21.2.0